refactor: autowired 애노테이션 제거
